I. #43 add a module to handle settings storage
